refactor: 가독성을 위한 패키지, 메서드, 상수, indent 분리
