+++ src/main/java/model/piece/state/Role.java
@@ -1,13 +1,12 @@
 package model.piece.state;
 
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
-import model.position.Position;
-import model.position.Route;
 import model.direction.MovingPattern;
 import model.piece.Color;
+import model.position.Position;
+import model.position.Route;
 
 public abstract class Role {
     protected Color color;
@@ -19,26 +18,24 @@ public abstract class Role {
     }
 
     public void checkSameCamp(Role role) {
-        if (this.color == role.color){
+        if (this.color == role.color) {
             throw new IllegalArgumentException("같은 진영의 기물이 목적 지점에 위치합니다.");
         }
     }
 
-    public boolean isOccupied(){
+    public boolean isOccupied() {
         return true;
     }
 
-    protected abstract Set<Route> possibleRoutes(Position position);
-
     public Route findRoute(Position source, Position destination) {
-        Set<Route> possibleRoutes = possibleRoutes(source);
-        for (Route route : possibleRoutes) {
-            if (route.contains(destination)) {
-                return route.subRoute(destination);
-            }
-        }
-        throw new IllegalArgumentException(" ");
+        return possibleRoutes(source).stream()
+                .filter(route -> route.contains(destination))
+                .map(route -> route.subRoute(destination))
+                .findAny()
+                .orElseThrow(() ->  new IllegalArgumentException("해당 기물이 이동할 수 없는 좌표입니다"));
     }
+
+    protected abstract Set<Route> possibleRoutes(Position position);
 
     protected Route getRoute(MovingPattern movingPattern, Position movedPosition) {
         List<Position> sequentialPositions = new ArrayList<>();
@@ -48,7 +45,6 @@ public abstract class Role {
         }
         return new Route(sequentialPositions);
     }
-
 
     public Color getColor() {
         return color;
