+++ src/main/java/view/OutputView.java
@@ -13,12 +13,13 @@ public class OutputView {
     private OutputView() {
     }
 
-    public static void printInitialGamePrompt(){
+    public static void printInitialGamePrompt() {
         System.out.println("체스 게임을 시작합니다.");
         System.out.println("> 게임 시작 : start");
         System.out.println("> 게임 종료 : end");
         System.out.println("> 게임 이동 : move source위치 target위치 - 예. move b2 b3");
     }
+
     public static void printChessBoard(List<PieceInfo> pieceInfos) {
         initializeChessBoard();
         pieceInfos.forEach(OutputView::placePieceOnBoard);
