+++ src/test/java/model/ChessBoardFactoryTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 
 import java.util.Map;
+import model.chessboard.ChessBoardFactory;
 import model.piece.PieceHolder;
 import model.piece.state.King;
 import model.piece.state.Pawn;
